requirements.txt
# AI Infrastructure Junior Engineer - Learning Repository Dependencies
# Python 3.11+ required

# ==============================================================================
# ML Frameworks
# ==============================================================================
torch>=2.0.0,<3.0.0
torchvision>=0.15.0
tensorflow>=2.13.0,<3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ==============================================================================
# Web Frameworks & API Development
# ==============================================================================
fastapi>=0.104.0
flask>=3.0.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# ==============================================================================
# MLOps Tools
# ==============================================================================
mlflow>=2.8.0
dvc>=3.30.0
great-expectations>=0.18.0
evidently>=0.4.0

# ==============================================================================
# Workflow Orchestration
# ==============================================================================
# Choose one based on your preference:
# apache-airflow>=2.7.0  # Uncomment for Airflow
prefect>=2.13.0          # Default choice for beginners

# ==============================================================================
# Monitoring & Logging
# ==============================================================================
prometheus-client>=0.18.0
python-json-logger>=2.0.0
loguru>=0.7.0

# ==============================================================================
# Testing & Code Quality
# ==============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
pylint>=3.0.0
isort>=5.12.0

# ==============================================================================
# Cloud SDKs
# ==============================================================================
boto3>=1.28.0              # AWS SDK
google-cloud-storage>=2.10.0  # GCP SDK
azure-storage-blob>=12.19.0   # Azure SDK

# ==============================================================================
# Database & Caching
# ==============================================================================
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
redis>=5.0.0
pymongo>=4.5.0

# ==============================================================================
# Container & Orchestration Tools
# ==============================================================================
docker>=6.1.0
kubernetes>=28.1.0

# ==============================================================================
# Data Processing
# ==============================================================================
pyarrow>=13.0.0
polars>=0.19.0
dask[complete]>=2023.10.0

# ==============================================================================
# Utilities
# ==============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
httpx>=0.25.0
click>=8.1.0
tqdm>=4.66.0
python-dateutil>=2.8.0
jsonschema>=4.19.0

# ==============================================================================
# Security
# ==============================================================================
cryptography>=41.0.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0

# ==============================================================================
# Model Serving
# ==============================================================================
bentoml>=1.1.0
tritonclient[all]>=2.40.0

# ==============================================================================
# Development Tools
# ==============================================================================
ipython>=8.16.0
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
pre-commit>=3.5.0

# ==============================================================================
# Documentation
# ==============================================================================
mkdocs>=1.5.0
mkdocs-material>=9.4.0
pdoc3>=0.10.0
